test: move imports helpers - charm_metrics_helpers - modules
